+++ Slicer/Utils/Exceptions/InternalException.cs
@@ -11,8 +11,11 @@ namespace Slicer.Utils.Exceptions
         public InternalException()
             : base() { }
 
-        public InternalException(string message)
-            : base(message) { }
+		public InternalException(string message)
+			: base(message) { }
+		
+		public InternalException(Dictionary<string, dynamic> data)
+			: base(data) { }
 
         public InternalException(string format, params object[] args)
             : base(string.Format(format, args)) { }